Fix: Enable build on MacOS renamed wifi.h and wifi.cpp to wifiHandling.h and wifiHandling.cpp to prevent IDE from including the local files as library.
Update to ESP Arduino 3
