+++ software/esp-firmware/config.cpp
@@ -25,7 +25,7 @@
 #include <ArduinoJson.h>
 
 #include "config.h"
-#include "wifi.h"
+#include "wifiHandling.h"
 #include "locoHandling.h"
 #include "throttleHandling.h"
