+++ software/esp-firmware/config.h
@@ -25,7 +25,7 @@
 #include <EEPROM.h>
 #include <stdbool.h>
 #include <stdint.h>
-#include "wifi.h"
+#include "wifiHandling.h"
 #include "locoHandling.h"
 
 // Filenames and field names on SPIFFS
